Some Level Block out Completed and Set up Sensors to work
I finished working out the large details of the level, but want to add more guards and alternate paths.

I also have the camera perform a sphere trace on now when the player is detected in the zone and make sure nothing is blocking the player, so if they are behind a wall all is well.