Player Character Updated to a Metahuman, SensorDetection Updated
I updated the character model to use a metahuman, as well as started to flesh out the level. Then I changed my sphere trace on the rotatative sensors to multi trace so that they can correctly detect if the player is the first thing hit, allowing the player to hide behind walls and objects.